Séparation QRCodeGen et stories
