+++ web/src/routes/staff/gen-qrcode/index.tsx
@@ -1,50 +1,6 @@
-import { createFileRoute, Link } from "@tanstack/react-router";
-import { useStandists } from "@/lib/hooks/useStandists.ts";
-import { Header } from "@/components/Header.tsx";
-import clsx from "clsx";
-import { imagePrefix, images } from "@/lib/images.ts";
+import { createFileRoute } from "@tanstack/react-router";
+import QRCodeGenArtistsList from "@/components/routes/staff/QRCodeGen/QRCodeGenArtistsList.tsx";
 
 export const Route = createFileRoute("/staff/gen-qrcode/")({
-  component: QRCodeGenList,
+  component: QRCodeGenArtistsList,
 });
-
-function QRCodeGenList() {
-  const { data: standistsList } = useStandists();
-
-  return (
-    <div className={"mb-4"}>
-      <Header>Usurper un artiste</Header>
-      {standistsList && (
-        <div className={"grid grid-cols-2 gap-4 overflow-x-scroll"}>
-          {standistsList.map((doc) => {
-            return (
-              <div
-                key={doc.userId}
-                className={clsx("flex flex-col items-center gap-2")}
-              >
-                {images[
-                  `${imagePrefix}${doc.image}` as keyof typeof images
-                ] && (
-                  <Link to={`/staff/gen-qrcode/${doc.userId}`}>
-                    <img
-                      src={images[`${imagePrefix}${doc.image}`]}
-                      alt={doc.name}
-                      className={"rounded-full w-32 border-8 border-secondary"}
-                    />
-                  </Link>
-                )}
-
-                <div className="text-center relative bg-secondary w-40 py-1 rounded-xl">
-                  <p>{doc.name}</p>
-                  <p>
-                    H{doc.hall} {doc.boothNumber}
-                  </p>
-                </div>
-              </div>
-            );
-          })}
-        </div>
-      )}
-    </div>
-  );
-}
